chore: Moved the access control service into the services folder
